chore(Highlight): Create separate Highlight component
