+++ components/Highlight/Highlight.js
@@ -1,0 +1,45 @@
+import React from 'react';
+import styled from 'styled-components';
+import { colors } from '../utils';
+
+const handleColorHover = color => {
+  switch (color) {
+    case 'colorPrimary':
+      return `background: ${colors.colorTertiary};`;
+    case 'colorSecondary':
+      return `background: ${colors.colorQuaternary};`;
+    case 'colorTertiary':
+      return `background: ${colors.colorPrimary};`;
+    case 'colorQuaternary':
+      return `background: ${colors.colorSecondary};`;
+    case 'colorDark':
+      return `background: ${colors.colorLight};`;
+    case 'colorLight':
+      return `background: ${colors.colorDark};`;
+    default:
+      return `background: ${colors.colorDark}; color: ${colors.colorLight}`;
+  }
+};
+
+const HighlightSpan = styled.span`
+  background: ${({ color }) => colors[color]};
+  border-radius: 4px;
+  font-size: 2rem;
+  font-weight: 800;
+  margin: 5px;
+  padding: 0 7px;
+  transition: 0.3s;
+  &:hover {
+    box-shadow: 4px 4px 0 0 ${colors.colorDark};
+    ${({ color }) => handleColorHover(color)};
+  }
+  @media screen and (max-width: 500px) {
+    font-size: 1.5rem;
+  }
+`;
+
+const Highlight = ({ children, color }) => {
+  return <HighlightSpan color={color}>{children}</HighlightSpan>;
+};
+
+export default Highlight;
