+++ pages/index.js
@@ -6,8 +6,9 @@ import Card from '../components/Card/Card';
 import About from './about';
 import Work from './work';
 import Blog from './blog';
+import Highlight from '../components/Highlight/Highlight';
 import styled from 'styled-components';
-import { colors, helpers } from './utils';
+import { colors, helpers } from '../components/utils';
 
 const IntroH1 = styled.h1`
   text-align: center;
@@ -17,41 +18,6 @@ const IntroH1 = styled.h1`
   margin: 0;
   @media screen and (max-width: 500px) {
     font-size: 2rem;
-  }
-`;
-
-const handleColorHover = color => {
-  switch (color) {
-    case 'colorPrimary':
-      return `background: ${colors.colorTertiary};`;
-    case 'colorSecondary':
-      return `background: ${colors.colorQuaternary};`;
-    case 'colorTertiary':
-      return `background: ${colors.colorPrimary};`;
-    case 'colorQuaternary':
-      return `background: ${colors.colorSecondary};`;
-    case 'colorDark':
-      return `background: ${colors.colorLight};`;
-    case 'colorLight':
-      return `background: ${colors.colorDark};`;
-    default:
-      return `background: ${colors.colorDark}; color: ${colors.colorLight}`;
-  }
-};
-
-const Highlight = styled.span`
-  background: ${({ color }) => colors[color]};
-  border-radius: 4px;
-  font-size: 2rem;
-  margin: 5px;
-  padding: 0 7px;
-  transition: 0.3s;
-  &:hover {
-    box-shadow: 4px 4px 0 0 ${colors.colorDark};
-    ${({ color }) => handleColorHover(color)};
-  }
-  @media screen and (max-width: 500px) {
-    font-size: 1.5rem;
   }
 `;
 
@@ -69,11 +35,8 @@ export default function Home({ allPostsData }) {
           <IntroH1>Hi! I'm Timothy Merritt</IntroH1>
           <h2 className='flexRow'>
             <Highlight color='colorPrimary'>developer</Highlight>
-            {' // '}
             <Highlight color='colorSecondary'>designer</Highlight>
-            {' // '}
             <Highlight color='colorTertiary'>writer</Highlight>
-            {' // '}
             <Highlight color='colorQuaternary'>musician</Highlight>*
           </h2>
           <sub>
